+++ examples/kyb-app/src/components/organisms/KYBView/flows/BaseFlow/useBaseFlow.ts
@@ -4,18 +4,26 @@ import { serializeBusinessData } from './helpers/serialize-business-data';
 import { useCallback, useMemo } from 'react';
 import { updateBusiness } from '@app/domains/business';
 import { serializeWorkflowRunData } from './helpers/serialize-workflow-run-data';
-import { runAndStartWorkflowRequest } from '@app/domains/workflows';
+import { runAndStartWorkflowRequest, updateWorkflow } from '@app/domains/workflows';
 import { WorkflowFlowData } from '@app/domains/workflows/flow-data.type';
 import { defaultFlowData } from '@app/domains/workflows/default-flow-data';
 
 export const useBaseFlow = () => {
   const handleFinish = useCallback(async (context: WorkflowFlowData) => {
-    const serializedBusinessPayload = serializeBusinessData(context, context.shared.businessId);
-    await updateBusiness(serializedBusinessPayload);
-
+    const isUpdate = Boolean(context.shared.workflowId);
     const serializedRunPayload = await serializeWorkflowRunData(context);
 
-    await runAndStartWorkflowRequest(serializedRunPayload);
+    if (isUpdate) {
+      await updateWorkflow({
+        workflowId: context.shared.workflowId,
+        payload: serializedRunPayload,
+      });
+    } else {
+      const serializedBusinessPayload = serializeBusinessData(context, context.shared.businessId);
+      await updateBusiness(serializedBusinessPayload);
+
+      await runAndStartWorkflowRequest(serializedRunPayload);
+    }
   }, []);
 
   const views = useMemo(() => attachStatusesToViews(kybViews, defaultFlowData), []);