+++ sdks/workflow-node-sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ballerine/workflow-node-sdk
 
+## 0.4.12
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.4.25
+
 ## 0.4.11
 
 ### Patch Changes