+++ services/workflows-service/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ballerine/workflows-service
 
+## 0.4.26
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.5.15
+  - @ballerine/workflow-core@0.4.25
+  - @ballerine/workflow-node-sdk@0.4.12
+
 ## 0.4.25
 
 ### Patch Changes